tech: Fix request

+++ db/migrate/20210115181509_create_order_item.rb
@@ -3,7 +3,7 @@ class CreateOrderItem < ActiveRecord::Migration[6.1]
     create_table :order_items do |t|
       t.references :order, null: false, foreign_key: true
       t.references :item, null: false, foreign_key: true
-      t.numeric :quantity
+      t.integer :quantity
       t.numeric :unit_price
       t.numeric :full_unit_price
